wasm: load state file support
previously it was only available on native

+++ src/apu.rs
@@ -1,6 +1,5 @@
 use modular_bitfield::prelude::*;
 
-#[cfg(not(target_arch = "wasm32"))]
 use savefile::prelude::*;
 
 static LENGTH_TABLE: [u8; 32] = [
@@ -24,8 +23,7 @@ static DMC_RATE_TABLE: [u16; 16] = [
 ];
 
 #[bitfield(bytes = 1)]
-#[derive(Debug, Clone, Default, Copy)]
-#[cfg_attr(not(target_arch = "wasm32"), derive(Savefile))]
+#[derive(Debug, Clone, Default, Copy, Savefile)]
 pub struct ApuStatus {
     pub enable_dmc: bool,
     pub enable_noise: bool,
@@ -37,8 +35,7 @@ pub struct ApuStatus {
     pub dmc_active: bool,
 }
 
-#[derive(Debug, Default, Clone)]
-#[cfg_attr(not(target_arch = "wasm32"), derive(Savefile))]
+#[derive(Debug, Default, Clone, Savefile)]
 pub struct Envelope {
     pub start: bool,
     pub disable: bool,
@@ -80,8 +77,7 @@ impl Envelope {
     }
 }
 
-#[derive(Debug, Default, Clone)]
-#[cfg_attr(not(target_arch = "wasm32"), derive(Savefile))]
+#[derive(Debug, Default, Clone, Savefile)]
 pub struct Sweep {
     pub enabled: bool,
     pub period: u8,
@@ -121,8 +117,7 @@ impl Sweep {
     }
 }
 
-#[derive(Debug, Default, Clone)]
-#[cfg_attr(not(target_arch = "wasm32"), derive(Savefile))]
+#[derive(Debug, Default, Clone, Savefile)]
 pub struct Pulse {
     pub enabled: bool,
     pub channel_idx: u8,
@@ -210,8 +205,7 @@ impl Pulse {
     }
 }
 
-#[derive(Debug, Default, Clone)]
-#[cfg_attr(not(target_arch = "wasm32"), derive(Savefile))]
+#[derive(Debug, Default, Clone, Savefile)]
 pub struct Triangle {
     pub enabled: bool,
     pub length_value: u8,
@@ -284,8 +278,7 @@ impl Triangle {
     }
 }
 
-#[derive(Debug, Default, Clone)]
-#[cfg_attr(not(target_arch = "wasm32"), derive(Savefile))]
+#[derive(Debug, Default, Clone, Savefile)]
 pub struct Noise {
     pub enabled: bool,
     pub length_value: u8,
@@ -356,8 +349,7 @@ impl Noise {
     }
 }
 
-#[derive(Debug, Default, Clone)]
-#[cfg_attr(not(target_arch = "wasm32"), derive(Savefile))]
+#[derive(Debug, Default, Clone, Savefile)]
 pub struct Dmc {
     pub enabled: bool,
     pub value: u8,
@@ -439,8 +431,7 @@ impl Dmc {
     }
 }
 
-#[derive(Debug, Clone, Copy)]
-#[cfg_attr(not(target_arch = "wasm32"), derive(Savefile))]
+#[derive(Debug, Clone, Copy, Savefile)]
 pub struct HighPassFilter {
     c: f32,
     prev_out: f32,
@@ -467,8 +458,7 @@ impl HighPassFilter {
     }
 }
 
-#[derive(Debug, Clone)]
-#[cfg_attr(not(target_arch = "wasm32"), derive(Savefile))]
+#[derive(Debug, Clone, Savefile)]
 pub struct Apu {
     pub status: ApuStatus,
     pub pulse1: Pulse,
@@ -476,11 +466,11 @@ pub struct Apu {
     pub triangle: Triangle,
     pub noise: Noise,
     pub dmc: Dmc,
-    pub frame_counter: usize,
+    pub frame_counter: u32,
     pub frame_mode: bool,
     pub frame_irq_inhibit: bool,
     pub frame_irq_pending: bool,
-    pub cycles: usize,
+    pub cycles: u32,
     hpf1: HighPassFilter,
     hpf2: HighPassFilter,
 }
